==> Development/1. Collecting entries/1. REALEC Accounting.ipynb <==
Initial import

from sklearn.model_selection import train_test_split
import pandas as pd
import tensorflow as tf
import tensorflow_hub as hub
from datetime import datetime

from tensorflow import keras
import os
import re

Download actual data from REALEC

  dataset = tf.keras.utils.get_file(
      fname="data.tar.gz", 
      origin="http://realec.org/ajax.cgi?action=downloadCollection&collection=%2F&protocol=1", 
      extract=True)

dataset

os.listdir("/root/.keras/datasets/data")

Move downloaded data to local playground

import shutil

shutil.copytree("/root/.keras/datasets/data","./data")

os.chdir("data")

Delete german files to clean the dataset


shutil.rmtree("de")

Check if any texts don't have a corresponding annotation file

loose_files = []

for root, dirs, files in os.walk('.'):
    for file in os.listdir(root):
        if file.endswith(".txt"):
            path = root+os.sep+file
            if not os.path.isfile(path[:-4]+".ann"):
              print(path,"does not have .ann!")
              loose_files.append(path)

And vice versa:

loose_annotations = []

for root, dirs, files in os.walk('.'):
    for file in os.listdir(root):
        if file.endswith(".ann"):
            path = root+os.sep+file
            if not os.path.isfile(path[:-4]+".txt"):
              print(path,"does not have .ann!")
              loose_annotations.append(path)

Delete the unpaired files:

for path in loose_files + loose_annotations:
  os.remove(path)
  print("Removed",path)

Now let's check the collection for duplicate text files. We won't search for duplicate annotations as two hypothetically empty annotations corresponding to two different ideal texts should not be interpreted as duplicates.

import sys
import os
import hashlib


def chunk_reader(fobj, chunk_size=1024):
    """Generator that reads a file in chunks of bytes"""
    while True:
        chunk = fobj.read(chunk_size)
        if not chunk:
            return
        yield chunk


def get_hash(filename, first_chunk_only=False, hash=hashlib.sha1):
    hashobj = hash()
    file_object = open(filename, 'rb')

    if first_chunk_only:
        hashobj.update(file_object.read(1024))
    else:
        for chunk in chunk_reader(file_object):
            hashobj.update(chunk)
    hashed = hashobj.digest()

    file_object.close()
    return hashed


def check_for_duplicates(paths, hash=hashlib.sha1):
    duplicate_pairs = []
    
    hashes_by_size = {}
    hashes_on_1k = {}
    hashes_full = {}

    for path in paths:
        for dirpath, dirnames, filenames in os.walk(path):
            for filename in filenames:
                full_path = os.path.join(dirpath, filename)
                try:
                    # if the target is a symlink (soft one), this will 
                    # dereference it - change the value to the actual target file
                    full_path = os.path.realpath(full_path)
                    file_size = os.path.getsize(full_path)
                except (OSError,):
                    # not accessible (permissions, etc) - pass on
                    continue

                duplicate = hashes_by_size.get(file_size)

                if duplicate:
                    hashes_by_size[file_size].append(full_path)
                else:
                    hashes_by_size[file_size] = []  # create the list for this file size
                    hashes_by_size[file_size].append(full_path)

    # For all files with the same file size, get their hash on the 1st 1024 bytes
    for __, files in hashes_by_size.items():
        if len(files) < 2:
            continue    # this file size is unique, no need to spend cpy cycles on it

        for filename in files:
            try:
                small_hash = get_hash(filename, first_chunk_only=True)
            except (OSError,):
                # the file access might've changed till the exec point got here 
                continue

            duplicate = hashes_on_1k.get(small_hash)
            if duplicate:
                hashes_on_1k[small_hash].append(filename)
            else:
                hashes_on_1k[small_hash] = []          # create the list for this 1k hash
                hashes_on_1k[small_hash].append(filename)

    # For all files with the hash on the 1st 1024 bytes, get their hash on the full file - collisions will be duplicates
    for __, files in hashes_on_1k.items():
        if len(files) < 2:
            continue    # this hash of fist 1k file bytes is unique, no need to spend cpy cycles on it

        for filename in files:
            try: 
                full_hash = get_hash(filename, first_chunk_only=False)
            except (OSError,):
                # the file access might've changed till the exec point got here 
                continue

            duplicate = hashes_full.get(full_hash)
            if duplicate:
                duplicate_pairs.append((filename, duplicate))
            else:
                hashes_full[full_hash] = filename

    return duplicate_pairs

duplicate_pairs = check_for_duplicates(".")

duplicate_txts = [pair for pair in duplicate_pairs if pair[0].endswith('.txt') and pair[1].endswith('.txt')]

import numpy as np

dup_ex = pd.DataFrame(duplicate_txts)
dup_ex.head()

import networkx as nx
G=nx.Graph()

G.add_edges_from(duplicate_txts)

conns = list(nx.connected_components(G))

conn_ex = [conn for conn in conns if len(conn) > 2]
conn_ex[:10] # example of detecting more than 2 duplicates

Now we will clear the dataset from duplicates. We will preserve the files with the most complete annotations (judging simply by file size)

import operator

duplicates_removed = 0

for duplicate_group in conns:
  duplicate_group = list(duplicate_group)
  duplicate_dict = {path: os.path.getsize(path[:-4]+".ann") for path in duplicate_group}
  maxval = max(duplicate_dict.items(), key=operator.itemgetter(1))[0]
  for path in duplicate_group:
    if path != maxval:
      # print (path, maxval)
      os.remove(path)
      os.remove(path[:-4]+".ann")
      duplicates_removed += 1

print("Duplicate files removed:", duplicates_removed)

Let's check if we have left any duplicates:

duplicate_pairs = check_for_duplicates(".")
duplicate_txts = [pair for pair in duplicate_pairs if pair[0].endswith('.txt') and pair[1].endswith('.txt')]
len(duplicate_txts)

That's it. Now we basically have a cleaned collection, free of junk files and duplicates. Here I prepare the collection for download as .tar.gz:

import tarfile

def make_tarfile(output_filename, source_dir):
    with tarfile.open(output_filename, "w:gz") as tar:
        tar.add(source_dir, arcname=os.path.basename(source_dir))

import datetime
import pytz

utc_offset = lambda offset: datetime.timezone(datetime.timedelta(seconds=offset))
dt = datetime.datetime.now()
tzone = pytz.timezone('Europe/Moscow')
dt = dt.astimezone(tzone)
fname = "realec_" + dt.strftime("%d%m%y_%H%M") + ".tar.gz"

os.chdir("..")
make_tarfile(fname, "data")

os.listdir(".")

One final step - we need to download the file from Colab's temporary workspace. Here's how it's done (uploading to your Google Drive):

# Install the PyDrive wrapper & import libraries.
# This only needs to be done once in a notebook.
!pip install -U -q PyDrive
from pydrive.auth import GoogleAuth
from pydrive.drive import GoogleDrive
from google.colab import auth
from oauth2client.client import GoogleCredentials

# Authenticate and create the PyDrive client.
# This only needs to be done once in a notebook.
auth.authenticate_user()
gauth = GoogleAuth()
gauth.credentials = GoogleCredentials.get_application_default()
drive = GoogleDrive(gauth)

# Create & upload a file.
uploaded = drive.CreateFile({'title': fname})
uploaded.SetContentFile(fname)
uploaded.Upload()
print('Uploaded file with ID {}'.format(uploaded.get('id')))
