==> Development/1. Collecting entries/2. REALEC Token preprocessing.ipynb <==
from sklearn.model_selection import train_test_split
import numpy as np
import pandas as pd
import tensorflow as tf
import tensorflow_hub as hub
from datetime import datetime

from tensorflow import keras
import os
import re

from google.colab import drive
drive.mount('/content/gdrive')

import shutil

drive_path = "Clean REALEC dumps/" #@param {type:"string"}
filename = "realec_110319_2315.tar.gz" #@param {type:"string"}


shutil.copy2('/content/gdrive/My Drive/'+drive_path+filename,'.')

import tarfile

tar = tarfile.open(filename)
tar.extractall()

# Dealing with corrupt strings in annotation files

##Atypical entries

Let's fetch all annotation entries into a DataFrame. Yes, I'm that crazy.

%%time

all_entries_path = []
all_entries_line = []


for root, dirs, files in os.walk('.'):
  for file in os.listdir(root):
    if file.endswith(".ann"):
      path = root+os.sep+file
      with open(path, 'r', encoding='utf-8') as annfile:
        lines = annfile.read().split('\n')
        for line in lines:
          if line != "":
            all_entries_path.append(path)
            all_entries_line.append(line)


all_entries = pd.DataFrame({
  'path': all_entries_path,
  'line': all_entries_line
})

all_entries.head()

all_entries

Back to work. Now let's figure out it there are any entries which don't follow the regular pattern.

%%time

atypical_entries = pd.DataFrame({
    'path': [],
    'line': []
})

for root, dirs, files in os.walk('.'):
  for file in os.listdir(root):
    if file.endswith(".ann"):
      path = root+os.sep+file
      with open(path, 'r', encoding='utf-8') as annfile:
        lines = annfile.read().split('\n')
        for line in lines:
          if line != "":
            if not re.search(r'^.[0-9]+\t', line):
              atypical_entries.loc[len(atypical_entries)] = [path, line]

atypical_entries

Let's just remove those from our general set so we don't have to worry about them afterwards. We'll also test for the time such an operation takes here.

%%time

all_entries = pd.concat([all_entries, atypical_entries]).drop_duplicates(keep=False)

Let's check if ```all_entries``` size has changed:


all_entries

Let's look which string prefixes we have:

linestarts = []

for line in list(all_entries["line"]):
  val = re.search(r'^(.*?)[0-9]', line).group(1)
  if val not in linestarts:
    linestarts.append(val)

linestarts

What are the `R` tags?

all_entries.loc[all_entries.line.str.contains('^R[0-9]'), :][:10]

So, what type of `R` tags causes do we have?

r_types = []

for line in list(all_entries["line"]):
  query = re.search(r'^R[0-9]+\t(.*?) ', line)
  if query:
    val = query.group(1)
  if val not in r_types:
    r_types.append(val)

r_types

What about the `A` non-delete ones?

a_types = []

for line in list(all_entries["line"]):
  query = re.search(r'^A[0-9]+\t(.*?) ', line)
  if query:
    val = query.group(1)
  if val not in a_types:
    a_types.append(val)

a_types

all_entries.loc[all_entries.line.str.contains('^A[0-9]+\tAddition'), :][:10]

What about `#`?

DASH_types = []

for line in list(all_entries["line"]):
  query = re.search(r'^\#[0-9]+\t(.*?) ', line)
  if query:
    val = query.group(1)
  if val not in DASH_types:
    DASH_types.append(val)

DASH_types

Based on these observations, I decided to further work with the `Delete` entries from all non-`A` tags and further narrowed my field of error tags to the following ones:

Allowed_types = ["Punctuation", "Spelling", "Capitalisation", "Grammar", "Determiners", "Articles", "Quantifiers", "Verbs", "Tense", "Tense_choice", "Tense_form", "Voice", "Modals", "Verb_pattern", "Intransitive", "Transitive", "Reflexive_verb", "Presentation", "Ambitransitive", "Two_in_a_row", "Verb_Inf", "Verb_Gerund", "Verb_Inf_Gerund", "Verb_Bare_Inf", "Verb_object_bare", "Restoration_alter", "Verb_part", "Get_part", "Complex_obj", "Verbal_idiom", "Prepositional_verb", "Dative", "Followed_by_a_clause", "that_clause", "if_whether_clause", "that_subj_clause", "it_conj_clause", "Participial_constr", "Infinitive_constr", "Gerund_phrase", "Nouns", "Countable_uncountable", "Prepositional_noun", "Possessive", "Noun_attribute", "Noun_inf", "Noun_number", "Prepositions", "Conjunctions", "Adjectives", "Comparative_adj", "Superlative_adj", "Prepositional_adjective", "Adj_as_collective", "Adverbs", "Comparative_adv", "Superlative_adv", "Prepositional_adv", "Numerals", "Pronouns", "Agreement_errors", "Word_order", "Standard", "Emphatic", "Cleft", "Interrogative", "Abs_comp_clause", "Exclamation", "Title_structure", "Note_structure", "Conditionals", "Attributes", "Relative_clause", "Defining", "Non_defining", "Coordinate", "Attr_participial", "Lack_par_constr", "Negation", "Comparative_constr", "Numerical", "Confusion_of_structures", "Vocabulary", "Word_choice", "lex_item_choice", "Often_confused", "lex_part_choice", "Absence_comp_colloc", "Redundant", "Derivation", "Formational_affixes", "Suffix", "Prefix", "Category_confusion", "Compound_word", "Discourse", "Ref_device", "Coherence", "Linking_device", "Inappropriate_register", "Absence_comp_sent", "Redundant_comp", "Absence_explanation"]

Allowed_types

# Further processing

Let's separate all `A Delete`, `#` and `T`-type tags from the rest:

ADelete_entries = all_entries.loc[all_entries.line.str.contains('^A[0-9]+\tDelete'), :]

ADelete_entries.head()

DASH_entries = all_entries.loc[all_entries.line.str.contains('^\#[0-9]'), :]

DASH_entries.head()

T_entries = all_entries.loc[all_entries.line.str.contains('^T[0-9]'), :]

T_entries.head()

`A`-entries should follow the pattern `A[any number]\tDelete␣T[any number]`. Let's check if we have any outliers!

ADelete_invalid = ADelete_entries.loc[~ADelete_entries.line.str.contains('^A[0-9]+\tDelete T[0-9]+$'), :]

ADelete_invalid

No invalid entries here. Moving to `#`, which should be `#[any number]\tAnnotatorNotes␣T[any number]\t[any text]`

DASH_invalid = DASH_entries.loc[~DASH_entries.line.str.contains('^#[0-9]+\tAnnotatorNotes T[0-9]+\t.*$'), :]

DASH_invalid

So there seem to be two cases: empty `AnnotatorNotes`, which we will assume to be en empty correction, and other stuff. Let's focus on the latter for now.

DASH_invalid_nonempty = DASH_entries.loc[~DASH_entries.line.str.contains('^#[0-9]+\tAnnotatorNotes T[0-9]+.*$'), :]

DASH_invalid_nonempty

We'll drop this ocurrence and add a trailing tabulation to the other irregular lines to unify our notation:

DASH_entries = pd.concat([DASH_entries, DASH_invalid_nonempty]).drop_duplicates(keep=False)

condition = ~DASH_entries.line.str.contains('^#[0-9]+\tAnnotatorNotes T[0-9]+\t.*$')

DASH_entries.loc[condition, 'line'] = DASH_entries.loc[condition, 'line']+'\t'

# DASH_entries.loc[condition, 'line'].apply(appendix) = DASH_entries.loc[condition, 'line'].apply(appendix)

DASH_new_invalid = DASH_entries.loc[~DASH_entries.line.str.contains('^#[0-9]+\tAnnotatorNotes T[0-9]+\t.*$'), :]

DASH_new_invalid

So, this narrows us down to `T`. These should be `T[any number]\t[error type]␣[any number]␣[any number]\t[any text]`. Let's test!

T_invalid = T_entries.loc[~T_entries.line.str.contains('^T[0-9]+\t[\w_]+ [0-9]+ [0-9]+\t.*?$'), :]

T_invalid

Right off the bat we can see that we have some double-staked entries here. Our best take is to divide these entries and treat them as independent tokens. Let's separate and process them:

T_double = T_entries.loc[T_entries.line.str.contains('^T[0-9]+\t[\w_]+ [0-9]+ [0-9]+;[0-9]+ [0-9]+\t.*?$'), :]

T_double

DEZip = zip(list(T_double.path), list(T_double.line))
double_entries = []
for e in DEZip:
  double_entries.append([e[0], e[1]])
double_entries

dedoubled_entries = []
for entry in double_entries:
  e = entry[1]
  head = re.search(r'^(T[0-9]+\t[\w_]+ )[0-9]+ [0-9]+;[0-9]+ [0-9]+\t.*?$', e).group(1)
  range1 = re.search(r'^T[0-9]+\t[\w_]+ ([0-9]+ [0-9]+);[0-9]+ [0-9]+\t.*?$', e).group(1)
  range2 = re.search(r'^T[0-9]+\t[\w_]+ [0-9]+ [0-9]+;([0-9]+ [0-9]+)\t.*?$', e).group(1)
  string = re.search(r'^T[0-9]+\t[\w_]+ [0-9]+ [0-9]+;[0-9]+ [0-9]+\t(.*?)$', e).group(1)
  len1 = int(range1.split(" ")[1])-int(range1.split(" ")[0])
  str1 = string[:len1]
  str2 = string[len1+1:]
  dedoubled_entries.append([entry[0], head+range1+'\t'+str1])
  dedoubled_entries.append([entry[0], head+range2+'\t'+str2])

dedoubled_entries

dedoubled = pd.DataFrame({
  'path': [],
  'line': []
})

for line in dedoubled_entries:
  dedoubled.loc[len(dedoubled)] = line

dedoubled

T_entries = pd.concat([T_entries, T_double]).drop_duplicates(keep=False)
T_entries = pd.concat([T_entries, dedoubled])

T_entries

Now let's get rid of all the tokens we will not be including in our analysis:

T_invalid = T_entries.loc[~T_entries.line.str.contains('^T[0-9]+\t('+'|'.join(Allowed_types)+') [0-9]+ [0-9]+'), :]

T_invalid

These are mostly parts of speech and suggestions. To hell with them:

T_entries = pd.concat([T_entries, T_invalid]).drop_duplicates(keep=False)

T_entries

Finally let's deal with the entries without suggestions. Like above, let's just find them and add a trailing tabulation:

T_nontabbed = T_entries.loc[~T_entries.line.str.contains('^T[0-9]+\t('+'|'.join(Allowed_types)+') [0-9]+ [0-9]+\t.*?$'), :]

T_nontabbed

condition = ~T_entries.line.str.contains('^T[0-9]+\t('+'|'.join(Allowed_types)+') [0-9]+ [0-9]+\t.*?$')

T_entries.loc[condition, 'line'] = T_entries.loc[condition, 'line']+'\t'

T_new_invalid = T_entries.loc[~T_entries.line.str.contains('^T[0-9]+\t('+'|'.join(Allowed_types)+') [0-9]+ [0-9]+\t.*?$'), :]

T_new_invalid

We've still got one monster left, as it seems. Let's drop it

T_entries = pd.concat([T_entries, T_new_invalid]).drop_duplicates(keep=False)

Let's reset indices of all resulting dataframes here:

ADelete_entries = ADelete_entries.reset_index(drop=True)
DASH_entries = DASH_entries.reset_index(drop=True)
T_entries = T_entries.reset_index(drop=True)

## Uploading the results

Wow. That was epic and took a looong time. We'd better save it into a json file to accelerate future load-ups.

%%time

jsonname = "DASH_entries.json" #@param {type:"string"}
with open(jsonname, 'w', encoding="utf-8") as outie:
    exec("outie.write("+jsonname[:-5]+".to_json())")

Whew, great! Let's upload this to Google Drive right now. 

# Install the PyDrive wrapper & import libraries.
# This only needs to be done once in a notebook.
!pip install -U -q PyDrive
from pydrive.auth import GoogleAuth
from pydrive.drive import GoogleDrive
from google.colab import auth
from oauth2client.client import GoogleCredentials

# Authenticate and create the PyDrive client.
# This only needs to be done once in a notebook.
auth.authenticate_user()
gauth = GoogleAuth()
gauth.credentials = GoogleCredentials.get_application_default()
drive = GoogleDrive(gauth)

# Create & upload a file.
uploaded = drive.CreateFile({'title': jsonname})
uploaded.SetContentFile(jsonname)
uploaded.Upload()
print('Uploaded file with ID {}'.format(uploaded.get('id')))
