==> Development/1. Collecting entries/3. REALEC Entry collection.ipynb <==
from sklearn.model_selection import train_test_split
import numpy as np
import pandas as pd
import tensorflow as tf
import tensorflow_hub as hub
from datetime import datetime

from tensorflow import keras
import os
import re

from google.colab import drive
drive.mount('/content/gdrive', force_remount=True)

import shutil

drive_path = ""
filenames = ["ADelete_entries.json", "DASH_entries.json", "T_entries.json"]

for filename in filenames:
  shutil.copy2('/content/gdrive/My Drive/'+drive_path+filename,'.')

os.listdir(".")

for jsonname in filenames:
  exec(jsonname[:-5]+" = pd.read_json(jsonname).reset_index(drop=True)")

T_entries[:10]

DASH_entries[:10]

ADelete_entries[:10]

Let's rearrange this all into one single database!

linesplit = list(T_entries.line.apply(lambda x: [x.split('\t')[0]] + x.split('\t')[1].split(' ') + [x.split('\t')[2]] ))

All_Entries = pd.DataFrame({
  'path': list(T_entries.path),
  'id': [x[0] for x in linesplit],
  'type': [x[1] for x in linesplit],
  'begin': [x[2] for x in linesplit],
  'end': [x[3] for x in linesplit],
  'substring': [x[4] for x in linesplit],
  'correction': ['' for _ in linesplit],
  'delete': [False for _ in linesplit]
})

All_Entries.head()

DASHZip = zip(list(DASH_entries.path), list(DASH_entries.line))
DASHList = []
for e in DASHZip:
  DASHList.append([e[0], e[1].split('\t')[1].split(' ')[1], e[1].split('\t')[2]])
DASHList[:10]

DASHDict = {e[0]+e[1]: e[2] for e in DASHList}

%%time

for i, row in All_Entries.iterrows():
  key = row[0]+row[1]
  Correction = ""
  if key in DASHDict:
    Correction = DASHDict[key]
  All_Entries.at[i,'correction'] = Correction

All_Entries.head()

ADZip = zip(list(ADelete_entries.path), list(ADelete_entries.line))
ADSet = set()
for e in ADZip:
  ADSet.add(e[0]+e[1].split('\t')[1].split(' ')[1])

%%time

for i, row in All_Entries.iterrows():
  key = row[0]+row[1]
  Delete = False
  if key in ADSet:
    All_Entries.at[i,'delete'] = True

All_Entries.head()

All_Entries.loc[(All_Entries["correction"] == "") & (All_Entries["delete"] == False)]

# Adding contexts

Let's get our data here

import shutil

drive_path = "Clean REALEC dumps/" #@param {type:"string"}
filename = "realec_110319_2315.tar.gz" #@param {type:"string"}


shutil.copy2('/content/gdrive/My Drive/'+drive_path+filename,'.')

import tarfile

tar = tarfile.open(filename)
tar.extractall()

Time to create another monster: a structure to contain all the texts we're interested in

%%time

Texts = list(set([x[:-4]+'.txt' for x in list(All_Entries["path"])]))
Text_Dict = {path: open(path, 'r', encoding='utf-8-sig').read() for path in Texts}

print(len(Text_Dict))

Now let's generate entries for our given tokens

import nltk

nltk.download("punkt")
nltk.download("perluniprops")

from nltk.tokenize import TweetTokenizer
tknzr = TweetTokenizer()

from nltk.tokenize.moses import MosesDetokenizer
detokenizer = MosesDetokenizer()

We have to change our substrings first to a character that never appears in the corpus to avoid further confusion. Let's check some things...

chr(8)

contain = []

for el in Text_Dict:
  if chr(8) in Text_Dict[el]:
    contain.append(Text_Dict[el])

print(len(contain))

Perfect.

def maskify(instr):
  replist = tknzr.tokenize(instr)
  for replacement in replist:
    instr = instr.replace(replacement, chr(8), 1)
  instr = instr.replace(chr(8)+chr(8), chr(8)+' '+chr(8))
  spacechars = chr(9)+chr(10)+chr(160)+" "
  intext = re.sub(r'[^'+spacechars+']'+chr(8), ' '+chr(8), instr)
  intext = re.sub(chr(8)+r'[^'+spacechars+']', chr(8)+' ', instr)
  if not re.search(chr(8)+r'['+spacechars+']', intext):
    intext = intext.replace(chr(8), chr(8)+' ')
  instr = instr.replace(chr(8), "[MASK]")
  return instr

Yes, this means that we're masking punctuation. So if we dropped a number of sentence delimeters here, we will just concatenate this contexts with adjacent sentences in future

maskify("Although, in my child time I wasn't a child,")

All_Entries['context'] = ""

All_Entries.begin = All_Entries.begin.astype(int)
All_Entries.end = All_Entries.end.astype(int)

%%time

for i, row in All_Entries.iterrows():
  fpath = row['path'][:-4]+'.txt'
  start = row['begin']
  end = row['end']
  raw = Text_Dict[fpath]
  substr = raw[start:end]
  raw = raw[:start] + maskify(substr) + raw[end:]
  raw = re.sub(r'\s', ' ', raw)
  raw = re.sub(r'( )+', ' ', raw)
  sentences = nltk.sent_tokenize(raw)
  for k in range(len(sentences)):
    if '[MASK]' in sentences[k]:
      if k < 1:
        All_Entries.at[i,'context'] = ' '.join(sentences[0:k+2])
      else:
        All_Entries.at[i,'context'] = ' '.join(sentences[k-1:k+2])
      break

All_Entries.head()

Let's deal with these strange cases where `begin` and `end` values differ but we're not given any `substring`. We should just get it manually!

for i, row in All_Entries.iterrows():
  if (row['end'] > row['begin']) and (row['substring'] == ""):
    fpath = row['path'][:-4]+'.txt'
    start = int(row['begin'])
    end = int(row['end'])
    raw = Text_Dict[fpath]
    All_Entries.at[i,'substring'] = raw[start:end]

Now all what is left is to also count the length of substrings in terms of words as measured by `nltk.TweetTokenizer`

def twttknzr_length(instr):
  return len(tknzr.tokenize(instr))

All_Entries['substr_words'] = All_Entries.apply(lambda x: twttknzr_length(x['substring']), axis=1)

All_Entries.loc[All_Entries['type'] == 'lex_part_choice']

## Uploading the results

Wow. That was epic and took a looong time. We'd better save it into a json file to accelerate future load-ups.

%%time

jsonname = "All_Entries.json" #@param {type:"string"}
with open(jsonname, 'w', encoding="utf-8") as outie:
    exec("outie.write("+jsonname[:-5]+".to_json())")

Whew, great! Let's upload this to Google Drive right now. 

# Install the PyDrive wrapper & import libraries.
# This only needs to be done once in a notebook.
!pip install -U -q PyDrive
from pydrive.auth import GoogleAuth
from pydrive.drive import GoogleDrive
from google.colab import auth
from oauth2client.client import GoogleCredentials

# Authenticate and create the PyDrive client.
# This only needs to be done once in a notebook.
auth.authenticate_user()
gauth = GoogleAuth()
gauth.credentials = GoogleCredentials.get_application_default()
drive = GoogleDrive(gauth)

# Create & upload a file.
uploaded = drive.CreateFile({'title': jsonname})
uploaded.SetContentFile(jsonname)
uploaded.Upload()
print('Uploaded file with ID {}'.format(uploaded.get('id')))
